Move validation functions to clkValidation.cpp
