+++ CMakeLists.txt
@@ -33,6 +33,7 @@ add_library(recombo_core STATIC
         src/bfacfMove.cpp
         src/edgePool.cpp
         src/clkRecombination.cpp
+        src/clkValidation.cpp
         src/mmc_config.cpp
         src/mmchain.cpp
         src/newsud.cpp
